+++ src/load/azext_load/data_plane/utils/utils_yaml_config.py
@@ -5,10 +5,12 @@
 
 import uuid
 
+from azext_load.data_plane.utils.constants import LoadTestConfigKeys
 from azext_load.data_plane.utils import validators
 from azure.cli.core.azclierror import (
     InvalidArgumentValueError,
 )
+from azure.cli.core.commands.parameters import get_subscription_locations
 
 from knack.log import get_logger
 
@@ -16,14 +18,14 @@ logger = get_logger(__name__)
 
 
 def yaml_parse_autostop_criteria(data):
-    if (isinstance(data["autoStop"], str)):
+    if (isinstance(data[LoadTestConfigKeys.AUTOSTOP], str)):
         # pylint: disable-next=protected-access
-        validators._validate_autostop_disable_configfile(data["autoStop"])
+        validators._validate_autostop_disable_configfile(data[LoadTestConfigKeys.AUTOSTOP])
         return {
             "autoStopDisabled": True,
         }
-    error_rate = data["autoStop"].get("errorPercentage")
-    time_window = data["autoStop"].get("timeWindow")
+    error_rate = data[LoadTestConfigKeys.AUTOSTOP].get(LoadTestConfigKeys.AUTOSTOP_ERROR_RATE)
+    time_window = data[LoadTestConfigKeys.AUTOSTOP].get(LoadTestConfigKeys.AUTOSTOP_ERROR_RATE_TIME_WINDOW)
     # pylint: disable-next=protected-access
     validators._validate_autostop_criteria_configfile(error_rate, time_window)
     autostop_criteria = {
@@ -36,15 +38,15 @@ def yaml_parse_autostop_criteria(data):
     return autostop_criteria
 
 
-def yaml_parse_splitcsv(data):
-    if not isinstance(data.get("splitAllCSVs"), bool):
+def _yaml_parse_splitcsv(data):
+    if not isinstance(data.get(LoadTestConfigKeys.SPLIT_CSV), bool):
         raise InvalidArgumentValueError(
             "Invalid value for splitAllCSVs. Allowed values are boolean true or false"
         )
-    return data.get("splitAllCSVs")
+    return data.get(LoadTestConfigKeys.SPLIT_CSV)
 
 
-def validate_failure_criteria(failure_criteria):
+def _validate_failure_criteria(failure_criteria):
     parts = failure_criteria.split("(")
     if len(parts) != 2:
         raise ValueError(f"Invalid failure criteria: {failure_criteria}")
@@ -57,15 +59,15 @@ def validate_failure_criteria(failure_criteria):
         raise ValueError(f"Invalid failure criteria: {failure_criteria}")
 
 
-def get_random_uuid():
+def _get_random_uuid():
     return str(uuid.uuid4())
 
 
 def yaml_parse_failure_criteria(data):
     passfail_criteria = {}
     passfail_criteria["passFailMetrics"] = {}
-    for items in data["failureCriteria"]:
-        metric_id = get_random_uuid()
+    for items in data[LoadTestConfigKeys.FAILURE_CRITERIA]:
+        metric_id = _get_random_uuid()
         # check if item is string or dict. if string then no name is provided
         name = None
         components = items
@@ -74,7 +76,7 @@ def yaml_parse_failure_criteria(data):
             components = list(items.values())[0]
         # validate failure criteria
         try:
-            validate_failure_criteria(components)
+            _validate_failure_criteria(components)
         except InvalidArgumentValueError as e:
             logger.error("Invalid failure criteria: %s", str(e))
         passfail_criteria["passFailMetrics"][metric_id] = {}
@@ -97,3 +99,41 @@ def yaml_parse_failure_criteria(data):
                 "requestName"
             ] = name
     return passfail_criteria
+
+
+def _parse_regionwise_loadtest_config(cmd, regionwise_loadtest_config):
+    logger.debug("Parsing regionwise load test configuration")
+    regional_load_test_config = []
+    subscription_locations = get_subscription_locations(cmd.cli_ctx)
+    location_names = [location.name for location in subscription_locations]
+    for region_load in regionwise_loadtest_config:
+        region_name = region_load.get(LoadTestConfigKeys.REGION)
+        if region_name is None or not isinstance(region_name, str):
+            raise InvalidArgumentValueError("Region name is required of type string")
+        if region_name.lower() not in location_names:
+            raise InvalidArgumentValueError(f"Invalid region: {region_name}. Expected Azure region")
+        engine_instances = region_load.get(LoadTestConfigKeys.ENGINE_INSTANCES)
+        if engine_instances is None or not isinstance(engine_instances, int):
+            raise InvalidArgumentValueError("Engine instances is required of type integer")
+        regional_load_test_config.append({"region": region_name.lower(), "engineInstances": engine_instances})
+    logger.debug("Successfully parsed regionwise load test configuration: %s", regional_load_test_config)
+    return regional_load_test_config
+
+
+def yaml_parse_loadtest_configuration(cmd, data):
+    load_test_configuration = {}
+    load_test_configuration["engineInstances"] = data.get(LoadTestConfigKeys.ENGINE_INSTANCES)
+    if data.get(LoadTestConfigKeys.REGIONAL_LOADTEST_CONFIG) is not None:
+        load_test_configuration["regionalLoadTestConfig"] = _parse_regionwise_loadtest_config(
+            cmd,
+            data.get(LoadTestConfigKeys.REGIONAL_LOADTEST_CONFIG)
+        )
+    # quick test and split csv not supported currently in CLI
+    load_test_configuration["quickStartTest"] = False
+    if data.get(LoadTestConfigKeys.QUICK_START):
+        logger.warning(
+            "Quick start test is not supported currently in CLI. Please use portal to run quick start test"
+        )
+    if data.get(LoadTestConfigKeys.SPLIT_CSV) is not None:
+        load_test_configuration["splitAllCSVs"] = _yaml_parse_splitcsv(data=data)
+    return load_test_configuration